# openapi.yml
openapi: 3.1.0
...
info:
  title: Faint Light
  version: 0.1.0
  summary: Fast, self-hosted plate solving with a nova.astrometry.net-compatible API.
  description: |
    Faint Light implements the HTTP contract of nova.astrometry.net so that
    imaging software such as NINA can use it as a drop-in local replacement,
    plus one extension of its own (`/api/skyview`).

    Conventions shared by every operation:

    * **No authentication.** `/api/login` accepts any API key, or none, and
      always succeeds. The session token it returns is never checked.
    * **Trailing slashes are optional.** Every `/api/...` and `/wcs_file/...`
      path is served both with and without a trailing slash; only the
      slash-less form is listed here.
    * **Errors** are JSON: `{"status": "error", "errormessage": "..."}` with
      an HTTP status of 400, 404, 422, 501, 503 or 504 as documented per
      operation.
    * **Jobs are in memory.** A submission and its job are created together
      on upload (nova creates the job later; clients handle both). A server
      restart forgets all jobs.
    * **Request bodies** are limited to 256 MiB.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  contact:
    name: Faint Light
    url: https://nightsky.observer/faint-light/
...
servers:
  - url: http://localhost:8000
    description: Default port (`FAINT_LIGHT_PORT`)
...
paths:
...
  /api/login:
    post:
      operationId: login
      summary: Obtain a session
      description: |
        Always succeeds. The body is read only for a friendly log line: a
        `request-json` form field (URL-encoded, as NINA sends it, or
        multipart, as `curl -F` sends it) or a bare JSON body, any of which
        may carry `{"apikey": "..."}`. A missing or malformed body still logs
        in.
      tags: [nova]
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/RequestJsonForm"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RequestJsonForm"
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /api/upload:
    post:
      operationId: upload
      summary: Submit an image for solving
      description: |
        Multipart upload of one image. Returns immediately with a submission
        id; poll `/api/submissions/{id}` for the job id, then `/api/jobs/{id}`
        until the status leaves `solving`, then read
        `/api/jobs/{id}/calibration`.

        Optional solve hints go in the `request-json` field as a JSON object
        (see `UploadRequestJson`). Without hints the solve is blind; Faint
        Light's warm start may still reuse the scale learned from earlier
        solves.

        Accepted image formats: FITS (BITPIX 8, 16, 32, -32 and -64, with
        BZERO/BSCALE applied, mono or multi-plane averaged to luma), JPEG,
        PNG and TIFF (8/16-bit, mono or RGB).
      tags: [nova]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  contentEncoding: binary
                  description: The image to solve.
                request-json:
                  $ref: "#/components/schemas/UploadRequestJson"
            encoding:
              file:
                contentType: image/fits, image/jpeg, image/png, image/tiff, application/octet-stream
              request-json:
                contentType: application/json
      responses:
        "200":
          description: Upload accepted and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: No `file` field, empty file, or unreadable multipart body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Solve queue full. The job is recorded as failed; retry later.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/submissions/{id}:
    get:
      operationId: submissionStatus
      summary: Submission status
      description: |
        Mirrors nova's submission record. Unlike nova, the `jobs` array is
        populated immediately after upload, so a client can skip straight to
        polling the job. `job_calibrations` lists `[job, job]` once the job has
        succeeded and is empty otherwise.
      tags: [nova]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Submission record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/jobs/{id}:
    get:
      operationId: jobStatus
      summary: Job status
      description: Poll this until `status` is `success` or `failure`.
      tags: [nova]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Current status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/jobs/{id}/calibration:
    get:
      operationId: jobCalibration
      summary: Solved calibration
      description: |
        The astrometric solution in nova's conventions. For JPEG, PNG and TIFF
        input the orientation and parity follow nova's convention for
        top-down raster images, so values agree with what nova would return
        for the same file.
      tags: [results]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Calibration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calibration"
        "404":
          description: No such job, or the job has not (yet) succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/jobs/{id}/info:
    get:
      operationId: jobInfo
      summary: Job summary
      description: |
        Status, original filename, start time and, once solved, the
        calibration. Object identification is not implemented, so
        `objects_in_field`, `machine_tags` and `tags` are always empty.
      tags: [results]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Job info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobInfo"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/jobs/{id}/objects_in_field:
    get:
      operationId: objectsInField
      summary: Objects in field (always empty)
      description: Present for contract compatibility. Object identification is not implemented; the list is always empty and the job id is not checked.
      tags: [results]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Empty list
          content:
            application/json:
              schema:
                type: object
                required: [objects_in_field]
                properties:
                  objects_in_field:
                    type: array
                    items: { type: string }
                    maxItems: 0
  /api/jobs/{id}/machine_tags:
    get:
      operationId: machineTags
      summary: Machine tags (always empty)
      description: Present for contract compatibility. Always empty; the job id is not checked.
      tags: [results]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Empty list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
  /api/jobs/{id}/tags:
    get:
      operationId: tags
      summary: Tags (always empty)
      description: Present for contract compatibility. Always empty; the job id is not checked.
      tags: [results]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Empty list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
  /api/jobs/{id}/annotations:
    get:
      operationId: annotations
      summary: Annotations (always empty)
      description: Present for contract compatibility. Always empty; the job id is not checked.
      tags: [results]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Empty list
          content:
            application/json:
              schema:
                type: object
                required: [annotations]
                properties:
                  annotations:
                    type: array
                    items: { type: object }
                    maxItems: 0
  /wcs_file/{id}:
    get:
      operationId: wcsFile
      summary: WCS as a FITS header
      description: |
        A minimal single-HDU FITS file (no data) carrying the solved TAN WCS:
        `CTYPE1/2`, `CRVAL1/2`, `CRPIX1/2` (1-based), the `CD` matrix,
        `IMAGEW` and `IMAGEH`. This is what astrometry.net's `solve-field`
        writes as `*.wcs`. No SIP distortion terms.
      tags: [results]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: FITS header
          content:
            application/fits:
              schema:
                type: string
                contentEncoding: binary
        "404":
          description: No such job, or the job has not succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /skyview/{id}/chart.svg:
    get:
      operationId: skyviewChart
      summary: All-sky chart for a skyview job
      description: The SVG rendered by `POST /api/skyview`, addressed by the `job` id it returned. This path has no trailing-slash variant.
      tags: [skyview]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: SVG chart
          content:
            image/svg+xml:
              schema:
                type: string
        "404":
          description: No chart for this job (only skyview jobs have one)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /new_fits_file/{id}:
    get:
      operationId: newFitsFile
      summary: Not available
      description: nova's re-exported FITS image. Not produced by Faint Light.
      tags: [stubs]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "404":
          $ref: "#/components/responses/NotFound"
  /annotated_display/{id}:
    get:
      operationId: annotatedDisplay
      summary: Not available
      description: nova's annotated preview image. Not produced by Faint Light.
      tags: [stubs]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "404":
          $ref: "#/components/responses/NotFound"
components:
...
  schemas:
...
    RequestJsonForm:
      type: object
      description: A form whose `request-json` field holds a JSON document as a string. This is how nova clients send every request body.
      properties:
        request-json:
          type: string
          contentMediaType: application/json
          examples: ['{"apikey": "anything"}']
    LoginRequest:
      type: object
      properties:
        apikey:
          type: string
          description: Ignored. Any value, or none, authenticates.
      additionalProperties: true
    LoginResponse:
      type: object
      required: [status, message, authenticated, session]
      properties:
        status:
          const: success
        message:
          type: string
          examples: ["authenticated user: anonymous"]
        authenticated:
          const: true
        session:
          type: string
          description: 32 hex characters. Never checked by the server.
          pattern: "^[0-9a-f]{32}$"
...